abstract network cli in ocrd_network
