+++ ocrd_network/ocrd_network/cli/__init__.py
@@ -1,0 +1,11 @@
+from .client import client_cli
+from .processing_server import processing_server_cli
+from .processing_worker import processing_worker_cli
+from .processor_server import processor_server_cli
+
+__all__ = [
+    'client_cli',
+    'processing_server_cli',
+    'processing_worker_cli',
+    'processor_server_cli'
+]
